% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKATMC.R
\name{SKATMC}
\alias{SKATMC}
\title{Sequence Kernel Associaiton Test for Multi-Categorical outcomes (SKAT-MC)}
\usage{
SKATMC(null.model, G, weights = NULL, weights.beta = c(1, 25))
}
\arguments{
\item{null.model}{an output object from the `SKATMC_Null_Model` function.}

\item{G}{a numeric genotype matrix with each row as a different individual
and each column as a separate variant Each genotype should be coded as 0, 1, 2,
for AA, Aa, aa, where A is a major allele and a is a minor allele. Missing values
are teporarily not allowed.}

\item{weights}{a numeric vector of weights for the weighted kernels.
It is \eqn{\sqrt(w)} in the SKAT paper. We do not recommend to use weights
when analyzing common variants.}

\item{weights.beta}{a numeric vector of parameters for the beta weights for the
weighted kernels with default (1, 25). If you want to use your own weights, please use the “weights”
parameter. It will be ignored if “weights” parameter is not null.}
}
\value{
Returns a vector of p values with each reference, and if more than one p value
generated, the function will automatically return a overall p value from the
Cauchy combination. For nominal data. Returns a single p value for ordinal data.
}
\description{
Sequence Kernel Associaiton Test for Multi-Categorical outcomes (SKAT-MC)
}
