% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKATMC.R
\name{SKATMC_Null_Model}
\alias{SKATMC_Null_Model}
\title{Get paramters of the NULL model}
\usage{
SKATMC_Null_Model(formula, data.type, ref = NULL, data = NULL)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the
NULL model to be fitted.}

\item{data.type}{an indicator of the outcome type, either 'nominal' or 'ordinal'.}

\item{ref}{an optional argument for specifying the reference group, only
valid for nominal data. It can be a scalar object or a vector containing levels
of the outcome. The default is `NULL` where every level of the outcome will
be tried as the reference.}

\item{data}{an optional data frame containing the variables in the model
(default=NULL).}
}
\value{
This function returns an object that has model parameters of
the NULL model of no association between genetic variables and outcomes.
After obtaining it, please use `SKATMC` function to conduct the association test.
}
\description{
Get paramters of the NULL model
}
